Added Movement_v2
There seems to be an issue with certain editions of VideoReader. Movement_v2 has been added to those who run into error: "Array formation and parentheses-style indexing with objects of class 'VideoReader' is not allowed.  Use objects of class 'VideoReader' only as scalars or use a cell array."

This code works exactly as before but the vidObjs are now cell arrays. I have not tested this on older matlab versions but this should work on R2021. Future updates may merge these together.